Number Calculator
Objective

25 years ago, a British computer magazine had a programming contest and
this was one of the puzzles.
There are a large number of 9 digit integers in the range 123456789 to
987654321 where each digit only appears once. What is the 100,000th
number in this sequence?

Example

The first number is 123456789, the second is 123456798, the third is
123456879 and so on. No digit can repeat so 122345675 is not a valid
number in this sequence.

The problem was "Write a program in that outputs the 100,000th number as
fast as possible.

Use any algorithm, except you cannot pre-calculate the answer and then
write a program that just prints the result. Your entry must calculate
the number!".